requirements.txt
rich>=13.0.0
prompt-toolkit>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
ollama>=0.1.0
boto3>=1.26.0